pyproject.toml
[tool.poetry]
name = "june-2023-pyton-drf333"
version = "0.1.0"
description = ""
authors = ["Vitaliy Dechyshyn <demchyshyn.v87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.1"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.8.0"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
